requirements.txt
beautifulsoup4
discord.py
google
humanize
lxml
matplotlib
requests
